Nasty hack using mergeProps to get mutators set up right
As discussed in STRPCONN-1, it seems (as determined experimentally)
that react-redux invokes the mapDispatchToProps function before the
constructor of the instance, so that it can't see the resources that
have been added at construction.  But the mergeProps function, when
defined, gets called after each mapStateToProps invocation. So now
all we do in Wrapper.mapDispatch is stash the dispatch function; and
the new mergeProps function uses it to do the real dispatch-mapping.

This works, and fixes an important preliminary part of STRPCONN-1.