Change format of state-keys
When there is a data-key, this is now prepended to the state-key,
searated by a hash. This is the same convention as is about to be
adopted in RESTResources, so that the rather unsatisfactory mockProps
function in there can correctly determine for _all_ resource types
whether or not they are related to a specific dataKey.